feat: register existing models in catalog during startup scan
The scanner now writes discovered models into the SQLite catalog as
Done jobs so the update checker can track them.  Also includes minor
cleanups: derive Default for Config/CivitaiConfig, &Path over &PathBuf
in public signatures, LORA→Lora variant rename, let-chain refactors,
and a Cargo.lock dependency refresh (libc 0.2.183, mac-notification-sys
0.6.11).

+++ src/daemon/queue.rs
@@ -29,7 +29,7 @@ pub async fn run(
     active: ActiveTasks,
     progress: ProgressMap,
 ) {
-    let max = config.daemon.max_concurrent_downloads.max(1) as usize;
+    let max = config.daemon.max_concurrent_downloads.max(1);
     let sem = Arc::new(Semaphore::new(max));
 
     loop {
